Create MediaPicker protocol to encapsulate the common methods for launching media related sources
